third_party/blink/web_tests/svg/dynamic-updates/SVGLineElement-dom-y1-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y1' attribute of the SVGLineElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS lineElement.getAttribute('y1') is "100"
PASS lineElement.getAttribute('y1') is "10"
PASS successfullyParsed is true

TEST COMPLETE
